+++ src/dotnet/Fable.Compiler/State.fs
@@ -27,26 +27,34 @@ type PathRef =
     | FilePath of string
     | NonFilePath of string
 
-type Project(projectOptions: FSharpProjectOptions, checkedProject: FSharpCheckProjectResults, fableCore: PathRef, isWatchCompile: bool) =
+type Project(projectOptions: FSharpProjectOptions, implFiles: Map<string, FSharpImplementationFileContents>,
+             errors: FSharpErrorInfo array,  fableCore: PathRef, isWatchCompile: bool) =
     let timestamp = DateTime.Now
     let projectFile = Path.normalizePath projectOptions.ProjectFileName
     let entities = ConcurrentDictionary<string, Fable.Entity>()
     let inlineExprs = ConcurrentDictionary<string, InlineExpr>()
     let normalizedFiles =
-        projectOptions.SourceFiles
-        |> Seq.map Path.normalizeFullPath
-        |> Set
+        let dic = Dictionary()
+        for f in projectOptions.SourceFiles do
+            dic.Add(Path.normalizeFullPath f, false)
+        dic
     let rootModules =
-        checkedProject.AssemblyContents.ImplementationFiles
-        |> Seq.map (fun file -> file.FileName, FSharp2Fable.Compiler.getRootModuleFullName file)
-        |> Map
+        implFiles
+        |> Map.filter (fun file _ -> not <| file.EndsWith("fsi"))
+        |> Map.map (fun _ file -> FSharp2Fable.Compiler.getRootModuleFullName file)
     member __.TimeStamp = timestamp
     member __.FableCore = fableCore
     member __.IsWatchCompile = isWatchCompile
-    member __.CheckedProject = checkedProject
+    member __.ImplementationFiles = implFiles
+    member __.Errors = errors
     member __.ProjectOptions = projectOptions
     member __.ProjectFile = projectFile
-    member __.NormalizedFilesSet = normalizedFiles
+    member __.ContainsFile(sourceFile) =
+        normalizedFiles.ContainsKey(sourceFile)
+    member __.IsCompiled(sourceFile) =
+        normalizedFiles.[sourceFile]
+    member __.MarkCompiled(sourceFile) =
+        normalizedFiles.[sourceFile] <- true
     interface ICompilerState with
         member this.ProjectFile = projectOptions.ProjectFileName
         member this.GetRootModule(fileName) =