+++ src/dotnet/Fable.Core/RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 1.3.0-beta-001
+
+* See dotnet-fable 1.3.0-beta-001 release notes
+
 ### 1.2.4
 
 * See dotnet-fable 1.2.4 release notes